Add control point patching pass.
This change adds a LLVM pass which replaces a call to a dummy control
(`yk_control_point`) point with a call to a dynamically generated
control point (`yk_new_control_point`) containing all of the necessary
logic to start and stop tracing, and to jump to machine code etc.

The logic is *very* simplified for now, but this should be a good
starting point.

In order for LLVM to not optimise away some of the control point logic
(e.g. the IOStruct) this pass has to run very late in the pipeline.

The pass is also only enabled when `--yk-patch-control-point` is passed.

See the comment at the top of ControlPoint.cpp for more details.

Co-authored-by: Edd Barrett <vext01@gmail.com>