Implement rudimentary trace stitching.
Currently when JITted code runs an iteration of a user loop, we
unconditionally fall back to the interpreter afterwards. This change
makes the control point call the JITted code repeatedly so that
consecutive iterations of a user loop can run JITted code.

This is could (and should) be more performant. See:
https://github.com/ykjit/yk/issues/442