Make the control point accept a pointer to a YkLocation.
Currently the JIT identifies a user-program location using a simple
integer, but it should use the `YkLocation` struct, which would allow
for more realistic control point behaviour (hot counts etc.).

This change modifies the type of the patched-in control point (and its
call-site) to use `YkLocation`.

Note that the control point is still very simple and uses the addresses
of YkLocations as an identifier. This is not how we intend this to work
in the long run: YkLocations should be freely moveable (see
`location.rs` for our preliminary intent from the old yk design).

Needless to say, for now we also bypass all of YkLocation's
multi-threaded synchronisation too.