Move most of the control point logic into to Rust code.
This change moves some of the (compile-time generated) control point
logic into Rust code. Functionally this change is a no-op.

The plan is to keep as little code in compile-time generated LLVM IR, as
it's quite laborious to write and experiment with. Also we are about to
grapple with concurrency, and doing that in compile-time generated IR
would be a nightmare.