Make control point work "by pointer".
Previously we'd pass the control point its live variables in a struct by
value, and it would return a (potentially mutated) struct containing the
new state.

Like this:

```
%new_vars = new_control_point(YkLocation *%loc, YkCtrlPontVars %vars)
```

This change makes it so that the struct is passed by pointer:

```
new_control_point(YkLocation *%loc, YkCtrlPontVars* %vars_ptr)
```

If the struct is mutated, it happens in-place. No new state is returned.

This is required so that we can pass opaque pointers to Rust for
Laurie's proposed `transition_location()` rework.