Force emission of a branch in FastISel.
This is similar to #29, but for FastISel. The change ensures that a LLVM
`br` always codegens a branch, even if fallthrough is possible.

We re-use --yk-no-fallthrough for this. To do so we move its definition
into a more generic location.

With this, the Lua interpreter runs until stop-gapping with the yk JIT.