Implement more precise liveness analysis.
Before we considered any Value whose definition dominates the call to
the control point as being a live variable. This is correct in the sense
that in includes all of the necessary live variables, but is obviously
an over-approximation (it includes things which have died in the lead-up
to the call to the control point).

Having lots of unnecessary variables in the live set makes the JIT IR
larger than necessary and imposes a higher memory footprint at the time
of calling the control point (where we pass a big struct of live
variables).

This change implements a precise liveness analysis in order to whittle
the live set down to the bare essentials.

We use the algorithm from Modern Compiler Implementation in Java (2nd
edition), by Andrew W. Appel.