Keeps regalloc from inserting instructions before stackmaps.
One problem with stackmaps is that the register allocator often inserts
instructions (typically spill reloads) inbetween a call and the stackmap
instruction. This not only messes with the offset reported by the stackmap
(which we previously fixed), but also changes the machine variables being
tracked. In other words, applying the offset fix, means we return to the right
location after deoptimisation, but doing so means the tracked variables don't
match anymore. To fix this properly, we need to make sure that no other
instructions may be inserted inbetween a call and the stackmap instruction.