Fix crashes in yk test suite.
Our previous attempt wasn't good enough: the ykllvm sync didn't merge.
This is because there are other registers besides SSP that we are unable
to deopt because they have no DWARF number (a requirement of a stackmap
register record).

This fixes the Yk test suite for now. We will have to pray that
restoring those niche registers is not required for correctness.