Rewrite indirect locations to register locations if they alias.
When a register and an indirect location alias, we attach any extra
locations to the register so that they are written to during deopt. This
is based on the assumption that stackmaps will always prioritise
tracking the register. There was no reason for us to assume this is
always the case, and in fact we have now found a bug where the indirect
location is tracked instead.

This commit fixes this issue, by checking during stackmap generation if
a tracked indirect aliases with a register. If that is the case we emit
the register (including the extra locations) into the stackmap in place
of the indirect.