Fix shadowstack alignments.
The code we used to align slots on the shadowstack had a small bug,
which meant we didn't use the correct alignment. This has been fixed
now. We also weren't aligning the entire stack according to the ABI,
e.g. on x64 this is typically 16 bytes. This has also been fixed.