Don't insert stackmaps into `yk_outline` functions.
A recent discussion brought to light the fact that there's no point in
emitting stackmaps into functions that are marked `yk_outline`. By
inserting stackmaps there, we pay a runtime cost (blocking AOT
optimisations) for no real reason.

One way of not having stackmaps in yk_outline functions, would be to
make all AOT IR fields for safepoints (in branch and call instructions)
optional, and then in yk_outline functions we simply insert a None.

This strategy, however, requires changes reaching down the pipeline and
the additional `Option`s would bloat the AOT IR since typically there
are lots of safepoints.

What this change does instead is to not emit IR for functions annotated
`yk_outline`. Instead we emit only a function declaration. This way, no
changes are required to the structure of the IR.

There is a knock on effect though: if we have no IR for a function, we
lose the ability, during outlining, to "skip over" any promotions, debug
strings, or idempotent promotes that that function may inject into the
trace.

To fix this we:
 - Prohibit the user from inserting yk_promote/yk_debug_str in any
   function marked `yk_outline`. This now results in a compiler error at
   AOT compile-time.
 - Turn off the automated ykllvm pass that inserts idempotent promotions
   for functions marked `yk_outline`.

Note that functions containing a control point are special exemptions.
Although such functions get marked `yk_outline` because they contain a
loop, by their very nature, we do trace them, and thus we must include
stackmaps.

Benchmarking suggests some benchmarks will speed up. CD should be about
2x as fast.