Enable backend optimisations for functions that can't be traced.
If a function is marked `yk_outline` (and doesn't contain a control
point), then we can never trace it, so *in theory* the optimisations
that would otherwise break the mapper could actually be used.

There's one small catch though: the trace builder still has to be able
to "outline over" `yk_outline` functions, and if the function has IR,
the trace builder will try to map the blocks within and make a total
mess of it, since the optimisations skewed the mappings.

This change therefore also prevents ykllvm from emitting AOT IR for
those functions, effectively making them appear to the trace builder as
foreign code, which is then outlined differently, but crucially
*without* the need for the mappings to be consistent.

One knock-on requirement is that no `yk_outline` function is permitted
to promote values any more, since without IR, we'd never know to consume
the promoted bytes.

(A lot of code here was borrowed from a simlar (but distinct) PR:
https://github.com/ykjit/ykllvm/pull/239)

When benchmarking CD and Richards, gives speedups in the realms of
10-20% ish.

Note also how CD spends more time in sys after this change, but runs
faster overall. Interesting.

```

1: /home/vext01/research/yklua/src/lua-before harness.lua cd 3 250
            Mean        Std.Dev.    Min         Median      Max
real        39.238      0.835       38.105      39.893      39.964
user        96.451      3.320       92.716      96.077      102.050
sys         162.216     11.851      139.544     166.891     171.612

1: /home/vext01/research/yklua/src/lua-after harness.lua cd 3 250
            Mean        Std.Dev.    Min         Median      Max
real        34.846      0.694       33.892      34.710      36.050
user        89.768      1.549       87.430      89.734      91.851
sys         173.440     14.941      152.802     173.968     195.308

1: /home/vext01/research/yklua/src/lua-before harness.lua richards 3 100
            Mean        Std.Dev.    Min         Median      Max
real        21.959      0.223       21.589      21.962      22.275
user        25.826      0.563       24.879      25.914      26.643
sys         3.761       0.203       3.406       3.787       4.004

1: /home/vext01/research/yklua/src/lua-after harness.lua richards 3 100
            Mean        Std.Dev.    Min         Median      Max
real        18.173      0.277       17.672      18.250      18.469
user        21.613      0.723       20.565      21.441      22.533
sys         3.387       1.087       2.385       3.103       5.377
```