Correct handling of dynamically sized Allocas
Sometimes an alloca doesn't have a known size at compile-time (e.g.
an array with a non-const size). When this happens, we can't pre-compute
its stack frame size so we don't know how big to make the shadow stack.

For now, when we encounter such cases we outline the function,
preventing it from being traced at all, this obviates the need for a
shadow stack at all.