Make shadow stack debuggable
Debugging the shadow stack inside gdb is hard because it moves all the
locals around breaking gdb assumptions. This patch makes it a little
easier to debug by creating a 'debug local' for each stack frame which
points to the frame's corresponding shadow stack offset. The
'shadowstack' pointer is then visible in gdb with 'info locals'.

This relies solely on LLVM's debug intrinsic library, so does not affect
code generation when the '-g' flag is not set.