Record start and end pointers for the shadow-stack buffer
In order to support VMs which use tracing garbage collection, we need to
provide a way for them to scan the shadow stack during marking. The
simplest way is export the start and end pointers for the shadow-stack
heap allocation as global pointers.

Currently, this treats the entire allocated shadow-stack buffer as a
potential rootset. A better (and more correct) approach would be to use
shadowstack_begin and shadowstack_0 as the set and ignore the unused
space. I haven't got this fully working yet, so for now this
overapproximation will do and helps aid debugging anyway.