Revert "Make shadow stack debuggable"
This reverts commit d61f22de8b2a68c2ce35590d37949277ea4d39f9, which
causes a test in yk to fail. Normally I would fix the yk test, but
an extra load is inserted as a result of this commit, and it's not
obvious to me that we would expect that. Probably this commit can be
rereverted, but it's better for us to get the backlogged yk PRs in while
we work out what's going on.