Various fixes related to the `FixStackmapsSpillReloads` pass.
The `FixStackmapsSpillReloads` moves stackmap instructions back to the
call instruction they belong to, after the register allocator has
inserted spill reloads inbetween.

When doing so, we didn't update `kill` flags or removed tracked
registers from the stackmap that were no longer valid. This has been
fixed and we now run the verifier on any changes we made to make sure
they are sane.