Don't insert basicblock tracer calls in yk_outline functions.
Before this change, our system runs nbody about 4.7x that of upstream
Lua.

First I confirmed that it's nothing to do with the quality of the traces
we are building (they are pretty much the same as we make with HWT, with
some inevitable variation due to the need for trace recorder calls).

perf tells me that __trace_basicblock() is the single biggest cost of
SWT (67% of nbody). I then double checked this by running nbody without
the JIT enabled, comparing performance with and without the SWT recorder
pass enabled (--yk-basicblock-tracer). With this flag enabled, we are
4.2x that of without!

So the main cost is undoubtedly recording blocks.

The first thing we can do is not insert calls to the block recorder in
functions marked yk_outline. That's what this change does.

This gives us an 84% speedup for SWT on nbody (on account of the JIT
being able to execute more traces -- the same number of traces are
recorded as before).

To my surprise nbody is now about 2x faster than upstream Lua running
nbody. A measurement I had to double check!