Fix a bug in indirect stackmap extra locations.
We record some memory spills as being the size of the register that we
are addressing relative to. This is incorrect.

Fixes db.lua in HWT debug mode, where the deopt checker was catching
this in the form of an overlapping deopt (because the size was twice
what it should be).