Add support for transitioning between optimised and unoptimised execution.
The software tracer has in yk has two major downsides: calls to the block
recording function are expensive; and in order to trace the interpreter we have
to disable many compile time optimisations. This commit addresses the former
and sets the foundation for the latter problem. It does so by creating two
versions of the interpreter, one optimised and one unoptimised, and transitions
between them depending on the current JIT state. When the interpreter is
running normally we use the optimised version; when we are tracing a hot loop,
we switch to the unoptimised version.

When module cloning is enabled, the original IR module functions are
split into two versions:

    Optimised functions (same as before)
    Unoptimised functions (with a `_yk_unopt` prefix)

In the optimised version calls to the block recording function go to
`__yk_trace_basicblock_dummy` which does nothing, as opposed to
`__yk_trace_basicblock` in the unoptimised version, which is responsible for
recording the basic blocks that were executed. Since
`__yk_trace_basicblock_dummy` calls are much cheaper, this significantly
reduces the overhead of the SWT tracing.