Don't consider "address taken" functions to automatically be traceable.
This is an evolving heuristic, and we're probably going to refine this
further but for now it seems that we're better off being conservative
with "should we consider this function as traceable or not". Just
because a function has its address taken does not mean that it will end
up in a trace; right now, at least in yklua, doing so makes performance
noticeably worse. Note: this is not a correctness issue. It is purely
related to performance.