Only restore the shadow stack pointer if a longjmp actually happened.
Currently we do the following around a setjmp:

```
load ssptr
setjmp(...)
restore ssptr
```

This pass changes the idiom to:

```
load ssptr
if (setjmp(...) != 0) {
    restore ssptr
}
```

Because if setjmp returns 0, then no longjmp occurred and we can avoid
the memory store.

Doesn't seem to make a great deal of performance difference, but we
agreed that this was probably worth including anyway in case it
compounds with other optimisations later.