Module cloning version 2.
This is a simpler cloning scheme that gets us some pretty good speedups.

I'm measuring about a 15% speedup on knucleotide using `multitime -n10`.

See the comment at the top of ModuleClone.cpp for how it works.

Could use a ykllvm test, but since we are in sprinting mode, that'll
come later.

More optimisations to come later hopefully.