Allow cloning of functions containing promotions or yk_debug_str.
We'd like to be able to have optmised versions of functions with
promotions or debug strings, but because optimised clones are nearly
always marked yk_outline, this would invalidate the assumption that
outlined functions don't contain promotions (how would we know what to
discard from the promote buffer during outlining if we have no IR for
the clone?).

This change clones functions with promotions or debug strings and then
retroactively removes the promotions/debug strings from the optimised
clone versions. We don't trace optimised clones, so promotions and debug
strings serve no purpose there anyway.

For yklua this means we can have an optmised version of precallC.

Seems to gives about a 5% speedup on knucleotide.