Run ISel and backend optimisations on optimised clones.
Before we added `optnone` to every function before instruction selection
and the backend ran.

This change makes it so we only add `optnone` to things that could be
traced.

This means that things marked yk_outline and optimised clone functions
do not get marked `optnone`, so in theory they should benefit from a
handful of optimisations that they didn't before.