Move module cloner earlier to allow more AOT inlining of optimised clones.
This pass moves the position of the module cloner pass much earlier in
the pipeline: immediately before the link-time inliner pass.

Optimised clones then have any `noinline` attribute removed, thus
allowing them to be inlined into other optimised clones if necessary.

(Why are we talking about `noinline`? Currently early in the pipeline we
mark functions containing loops noinline and yk_outline so that they
don't end up getting traced, but this isn't required for optimised
clones which will never be traced).

Note that the new position of the module clone pass means it would only
take effect if the -O optimisation level is >= 2.

Gives about an 8% speedup on storage. According the earlier
investigation this is a benchmark that still spends quite a lot of time
in optimised clones.

(Fix CI while for new Debian while we are here)