Idempotent functions version 2.
There's no need to serialise calls to the idempotent recorder. Instead
we emit yk AOT IR as if they don't exist. The JIT runtime can infer that
the call was idempotent based on the callee.

This is quite a large change, because in order to "leave out" the
recorder calls we have to stitch up the gap we'd otherwise leave in the
SSA data flow.

We pretend that the result of the recorder call was actually created by
the call to the idempotent function. This breaks an assumption the
serialiser was making: that the basic block index of a value in the LLVM
IR is the same as the basic block index of the corresponding value in
the yk IR, so we have to work around this.