Add "shimming" to improve performance of AOT code when we are not tracing.
For functions that could be traced, this pass replaces calls to Yk
instrumented functions with calls to a little "shim" functions.

Each shim is basically of the form:

```
__yk_shim_f(arg0, ...) {
  if (__yk_is_thread_tracing) {
    return call f(arg0, ...);
  } else {
    return call __yk_opt_f(arg0, ...);
  }
}
```

Where `__yk_is_thread_tracing` is a native TLS variable defined by the
JIT runtime.

This means that if the system is not tracing, it can swiftly enter
uninstrumented AOT code at the next shimmable call and avoid the immense
cost of calling __yk_trace_basicblock() frequently for no reason.