Light(er)weight SWT.
This changes the software tracing pass so that we do the "is the thread
tracing?" check inline in the interpreter AOT code, instead of calling
out to Rust every time to do the check.

The AOT IR serialiser is then modified so as to give the JIT runtime the
IR as though the checks didn't exist. This is because:

 - a) we don't ever want to trace the checks, so they are just IR bloat.

 - b) it would be laborious to filter out the checks when building the
      trace, and the more we can do statically the better.

Measurement suggests small speedups in the order of a few percent points
for a handful of benchmarks, but this is (hopefully) only the beginning
of what can be done. There are many optimisations we plan to do which
*should* compound.