For the trace recorder, use a wrapper function with the PreserveAll ABI.
We use a wrapper so we can use an ABI that disrupts register allocation
in the interpreter code to a lesser extent. The wrapper is required
because the actual recorder function is defined in the JIT runtime's
shared object and we can't change the ABI for library calls.

Benchmarking suggests some decent speedup for a handful of benchmarks,
mostly in the 2%-10% range.